♻️ Refactor : 스토어 상세페이지 리팩토링 & 지성님 코멘트 반영
