+++ src/domains/product/components/ProductTag.tsx
@@ -2,11 +2,11 @@
 
 import { transformTagToKr } from '@/commons/constants/transfromTag';
 
-interface TagContainerProps {
+interface ProductTagProps {
   tag: string;
 }
 
-const TagContainer = ({ tag }: TagContainerProps) => (
+const ProductTag = ({ tag }: ProductTagProps) => (
   <div className="h-[21px] bg-white rounded border border-solid border-gray-200 justify-center items-center gap-1 inline-flex">
     <span className="text-gray-600 px-[6px] py-[2px] leading-150 text-11 font-normal">
       {transformTagToKr(tag)}
@@ -14,4 +14,4 @@ const TagContainer = ({ tag }: TagContainerProps) => (
   </div>
 );
 
-export default TagContainer;
+export default ProductTag;
